pyproject.toml
[project]
name = "smart-recipe-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "djangorestframework>=3.16.0",
    "psycopg2-binary>=2.9.10",
]
